Message: raise SMS text message length limit to 160 chars
Technically, we should be able to send SMS text messages up to 160 7-bit
ASCII characters, rather than the 140 8-bit character limit.  Not sure what
the limitations are going to be at the message delivery side, but we'll
have to deal with that when we come to it.

Note: we're assuming all SMS data currently being inserted into the database
can be converted to valid 7-bit ASCII.  No validation is currently being
done to double-check this.